API refactoring and bug fixes
