+++ frontend/src/app/data/objects/user.ts
@@ -1,5 +1,5 @@
 export interface User {
-    id?: number;  // Unique ID of the user
+    id: number;  // Unique ID of the user
     accessToken: string;  // Access token for authentication
     email: string;  // Email (used as username)
     name: string;  // Display name
